+++ github_repository/main.tf
@@ -11,6 +11,12 @@ resource "github_repository" "repo" {
 
   is_template = var.is_template
 
+  # parent template repo
+  template {
+    owner      = var.template_repo_owner
+    repository = var.template_repo_name
+  }
+
   # merge-related options
   allow_merge_commit     = var.allow_merge_commit
   allow_squash_merge     = var.allow_squash_merge